requirements.txt
sys
xml.etree.ElementTree
os
math
nltk.corpus